refactor(shopping-store): decouple shopping store domain from API enums
