+++ commerce/shopping-store/src/main/java/ru/yandex/practicum/smarthometech/commerce/store/domain/enums/QuantityState.java
@@ -1,0 +1,8 @@
+package ru.yandex.practicum.smarthometech.commerce.store.domain.enums;
+
+public enum QuantityState {
+    ENDED,
+    FEW,
+    ENOUGH,
+    MANY
+}
