+++ commerce/shopping-store/src/main/java/ru/yandex/practicum/smarthometech/commerce/store/domain/repository/ProductRepository.java
@@ -4,8 +4,8 @@ import java.util.Optional;
 import java.util.UUID;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
-import ru.yandex.practicum.smarthometech.commerce.api.dto.store.ProductCategory;
-import ru.yandex.practicum.smarthometech.commerce.api.dto.store.ProductState;
+import ru.yandex.practicum.smarthometech.commerce.store.domain.enums.ProductCategory;
+import ru.yandex.practicum.smarthometech.commerce.store.domain.enums.ProductState;
 import ru.yandex.practicum.smarthometech.commerce.store.domain.entity.Product;
 
 public interface ProductRepository {
